+++ frontend/src/components/Preview/Preview.js
@@ -1,13 +1,11 @@
-import React from 'react';
 import { FONT_COLOR, PATH } from '../../constants';
-import { BUTTON_BACKGROUND_TYPE } from '../Button/Button.styles';
-import Button from '../Button/Button';
-import Frame from '../Frame/Frame';
 import { Title, Container, ButtonContainer } from './Preview.styles';
 import PropTypes from 'prop-types';
 import PreviewList from '../PreviewList/PreviewList';
 import { useHistory } from 'react-router-dom';
 import { RightArrowIcon } from '../../assets/icons';
+import { Button, Frame } from '../common';
+import { BUTTON_BACKGROUND_TYPE } from '../common/Button/Button.styles';
 
 const Preview = ({ title }) => {
   const history = useHistory();
