+++ frontend/src/components/PreviewList/PreviewList.js
@@ -1,14 +1,14 @@
 import PreviewItem from '../PreviewItem/PreviewItem';
 import { Container, Error } from './PreviewList.styles';
 import { useFetch } from '../../hooks';
-import { requestgetAllReviewList } from '../../requests';
+import { requestGetAllReviewList } from '../../requests';
 import { ERROR_MESSAGE, PATH } from '../../constants';
 import { useHistory } from 'react-router-dom';
 
 const PreviewList = () => {
   const history = useHistory();
 
-  const { response: reviewList, error: reviewError } = useFetch([], requestgetAllReviewList);
+  const { response: reviewList, error: reviewError } = useFetch([], requestGetAllReviewList);
 
   const goReviewDetailPage = (id) => {
     history.push(`${PATH.REVIEW}/${id}`);
