+++ frontend/src/components/ReviewWritingModal/ReviewWritingModal.js
@@ -3,13 +3,11 @@ import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
 import { useSnackbar } from 'notistack';
 import { RESPONSE_STATE, SNACKBAR_MESSAGE, VACCINATION } from '../../constants';
-import Modal from '../Modal/Modal';
-import Selection from '../Selection/Selection';
-import Button from '../Button/Button';
 import { Container, TextArea, ButtonWrapper, buttonStyles } from './ReviewWritingModal.styles';
-import { BUTTON_SIZE_TYPE } from '../Button/Button.styles';
+import { BUTTON_SIZE_TYPE } from '../common/Button/Button.styles';
 import { postReviewAsync } from '../../service';
 import { findKey } from '../../utils';
+import { Button, Modal, Selection } from '../common';
 
 const ReviewWritingModal = ({ getReviewList, onClickClose }) => {
   const accessToken = useSelector((state) => state.authReducer?.accessToken);
