+++ frontend/src/components/common/Avatar/Avatar.stories.js
