+++ frontend/src/components/common/Avatar/Avatar.styles.js
