+++ frontend/src/components/common/BaseLayout/BaseLayout.styles.js
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { THEME_COLOR } from '../../constants';
+import { THEME_COLOR } from '../../../constants';
 
 const Container = styled.div`
   display: flex;
