+++ frontend/src/components/common/Button/Button.styles.js
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { PALETTE } from '../../constants';
+import { PALETTE } from '../../../constants';
 
 const BUTTON_SIZE_TYPE = Object.freeze({
   SMALL: 'SMALL',
