+++ frontend/src/components/common/Frame/Frame.stories.js
