+++ frontend/src/components/common/Frame/Frame.styles.js
