+++ frontend/src/components/common/Input/Input.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import PropTypes from 'prop-types';
 import { Label, LabelText, InputElement } from './Input.styles';
