+++ frontend/src/components/common/Modal/Modal.styles.js
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import { FONT_COLOR } from '../../constants';
+import { FONT_COLOR } from '../../../constants';
 
 const Container = styled.div`
   position: fixed;
