+++ frontend/src/components/common/index.js
@@ -1,0 +1,12 @@
+import Button from './Button/Button';
+import Frame from './Frame/Frame';
+import Input from './Input/Input';
+import Label from './Label/Label';
+import Modal from './Modal/Modal';
+import Selection from './Selection/Selection';
+import Tabs from './Tabs/Tabs';
+import Avatar from './Avatar/Avatar';
+import BaseLayout from './BaseLayout/BaseLayout';
+import SideBar from './SideBar/SideBar';
+
+export { Avatar, BaseLayout, Button, Frame, Input, Label, Modal, Selection, Tabs, SideBar };
