+++ frontend/src/components/index.js
@@ -1,0 +1,8 @@
+import Preview from './Preview/Preview';
+import PreviewItem from './PreviewItem/PreviewItem';
+import PreviewList from './PreviewList/PreviewList';
+import ReviewItem from './ReviewItem/ReviewItem';
+import ReviewWritingModal from './ReviewWritingModal/ReviewWritingModal';
+import VaccinationState from './VaccinationState/VaccinationState';
+
+export { Preview, PreviewItem, PreviewList, ReviewItem, ReviewWritingModal, VaccinationState };
